*Now it checks to see if the Google Analytics ID is empty rather than just ''.
*The function meant to use relative URLs whenever possible was fixed.
The problem with it is hard to explain, but I'll try:
If WP was installed in a subfolder that wasn't the root of the domain
(e.g. example.com/wordpress instead of example.com/ ), the home_url()
would be '/' (so even if WP was installed on example.com/wordpress the
home_url() would be example.com ).