Lots of changes, full class build of the SMOF framework. All main items from the Wordpress codebase as well.
Every option SMOF has as well as wordpress base has a class. Each
element of control will be custom in order to allow for tooltips and
other options. When unsure, the new wordpress default is utilized for
features. Many of the core items work, but some do not. Still a few
days left of work to get this perfect.