Update to the way things are saved. This fully integrates the less compiler. See notes to modify a core Wordpress file in order for this to function. Core file will be fixed in 3.6.
See http://core.trac.wordpress.org/ticket/24401 for exact code and/or
patch you should manually add. This will be in the Wordpress 3.6
release. It enables a hook that allows us to do functions before and
after a customizer save is initialized.

Any time a variable marked with less is saved, the CSS generator will
rerun for the front end.

Again, you must add that small single line of text for this to work. It
will require the use of Wordpress 3.6+ when the theme goes live.

Also writing the less styles to a file for the customizer.