Added Bower support, improved Gruntfile.js
This includes bower support. The files will be installed in
assets/vendor.
"less/app.less" contains the list of all bootstrap files, to be easier
to disable components that you don't need.
Gruntfile.js will get the name of the package from package.json.
grunt-contrib-uglify  was updated to latest version. Gruntfile was
changed to support it.