Fix HTML5 Validator Error
Prior to this commit the W3 HTML5 validator (http://validator.w3.org/) returns a “Bad value X-UA-Compatible for attribute http-equiv on element meta.”.

This commit moves this meta attribute to a HTTP header

Yes, this is nitpicking - but it is annoying to get a red flag when everything else is green. There have been several discussions about this over at the HTML5 boilerplate project, but that is about a IE conditional, that does not work.