Update ConditionalTagCheck and usage docs
When passing an argument to a function for ConditionalTagCheck, the
argument should *not* always be an array. The function will be called
wit the arguments as is and not modified. So if the function takes a
string argument, keep it as a string. If it takes an array, use an
array.