Provides cleaner names for page templates.
As detailed in issue #1936 (https://github.com/roots/sage/issues/1936),
recent changes to this filter break soberwp/controller in the specific
case of providing controller data to Blades for custom page templates
because WordPress stores the entire template filename (with path) as the
'name' of the template. This information is then duplicated when it runs
through Sage's `App\filter_templates()` function.

This commit simply checks for those redundant items and removes them if
they exist, providing nice, clean names for templates, and allowing
soberwp/controller to work properly.