Handling of cases involving complex Bootstrap rules
This also includes a revision to the new wide/full filter which caused issues with blocks that had no alignment set.

Theme variables have been slightly reworked to use maps. There is a new helper function to make working with the maps a little less raw.

There is a new mixin `bootstrap-else()` that takes two parameters. Either param may be an scss list or singular. This mixin does a check against the theme variable `$wp-blocks-bootstrap-enabled`. If true it will apply an @extend with the value of the first param, or a set of @extends if a list was specified. If false the second param will be used in the same way.

The Bootstrap button classes are too substantial to replicate. I think this is a reasonable accomodation and results in no extra weight for builds thanks to the usage of % and @extend.