Add src and public mix path helpers
Since Mix uses the `webpack.mix.js` path as the webpack context and
doesn’t let us change the context, having a helper can make it easier to
add new assets. It also provides a single source of truth for the assets
source directory (one less thing to _mix_ up — pun intended).

Also, this adds a `public` function which just makes it easy for setting
the `from` path in the `copyDirectory` task.

I figure by including the definitions of these functions within the
`webpack.mix.js` file itself it makes it clear to Sage users that they
are not a part of Mix.