Move built assets to storage/theme/assets
I think it is easier to have 1 directory (storage) for generated files
instead of 2 (storage and dist).

As far as semantics go, `public` — Laravel’s default for built files —
doesn’t make sense for Sage since it isn’t the only public directory in
Sage. `dist` is a bit better, however, I think it’s still a bit
misleading since it’s not the only theme folder which is distributed.
`storage` makes sense, since it a place for the theme to store files it
creates.

Obviously, `storage/theme/assets` is rather long to write out, but that
could be remedied by using a `mix.config.publicPath` helper within
`webpack.mix.js`.