Replaces setup.php with a proper service provider
`ThemeBaseServiceProvider.php` probably belongs in Acorn as a lower level utility. If this change is found to be desirable it could be moved, leaving only `ThemeServiceProvider` (which now extends it) in `app/Providers`.

I think this PR's declarative style is an improvement for developer workflows and a net gain in terms of maintainability. It is otherwise identical in terms of its feature-set to the existing implementation.