Simplify Mix Browsersync API
Since upstream changes were merged which make the default files to watch
use the publicPath (JeffreyWay/laravel-mix#2091). One caveat is that
those the default files to watch doesn’t include `config/*.php`, but the
impact for us would be low since those files probably wouldn’t be
changing a ton. It could be something we make another upstream PR for
since it could be beneficial to the Laravel community as well.