Use collect() instead of a view composer for alert
Tailwind is about staying in the markup as much as possible. This seems
like a better alternative than a View Composer or a switch statement,
since it keeps the class names in the markup, doesn’t mangle the HTML,
and let’s use only leave the Composers out of th purgecss content paths.