Set browserslist explicitly and exclude IE 11
A few tools we’re already using (Babel, Autoprefixer, Stylelint, ESlint)
use Browserslist under the hood for their target browsers. I think
explicitly setting it would be helpful for developers using Sage since
it will encourage them to configure it themselves instead of relying on
the invisible default setting ("defaults").

I’m using the `browserslist` key in `package.json` as the method since
[it’s recommended](https://github.com/browserslist/browserslist#queries),
however, we can always split this out to its own config file
(`.browserslistrc`). I think the former is better since `package.json`
is a heavily used file and it will make it very clear to the developer
which browsers they are supporting. A file that lists dependencies also
makes sense to provide a list of compile targets.

The query I’ve set also excludes IE 11 since I believe we were already
doing so. You can see the list of supported browsers this includes:

https://browsersl.ist/?q=defaults%2C+not+IE+11

Something to note is that since we aren’t using the JS API for
browserslist we do not need to include it as a (dev-)dependency for
Sage.

A downside of explicitly adding this is that it may cause an increase in
PR in regards to what the defaults should be set to.