Use version_compare in order
I didn't run into an issue until today, which is weird, but I'm getting a nonstop "You must be using PHP 7.2 or greater." even though I'm on 7.2.28. I looked into it and found that this version_compare is being called backwards! https://www.php.net/manual/en/function.version-compare.php. Also, I threw in a >= comparator because of the wording "5.2 or greater," "7.2 or greater"