Get theme as disk
The theme disk is too difficult to get to. Fix it!

I don't really intend to use the URL thing with it, it's just to have easy access to the Storage facade to write files when a folder isn't there. Maybe there's a better way to do this, but I've been using it and haven't found a downside; maybe someone has a better way!